Create locale plugin and apply to app user
