+++ packages/mongoose-plugins/src/index.js
@@ -2,6 +2,7 @@ const applicationPlugin = require('./application-plugin');
 const domainValidator = require('./domain-validator');
 const emailPlugin = require('./email-plugin');
 const emailValidator = require('./email-validator');
+const localePlugin = require('./locale-plugin');
 const ipValidator = require('./ip-validator');
 const orgRolePlugin = require('./org-role-plugin');
 const orgRoles = require('./org-roles');
@@ -12,6 +13,7 @@ const referencePlugin = require('./reference-plugin');
 module.exports = {
   applicationPlugin,
   emailPlugin,
+  localePlugin,
   orgRolePlugin,
   organizationPlugin,
   paginablePlugin,
